% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChineseUI.R
\name{ChineseUI}
\alias{ChineseUI}
\title{ChineseUI}
\usage{
ChineseUI()
}
\description{
Launch the Statistics Toolbox (Chinese Version) Graphical User Interface
}
\details{
This function will automatically launch the Statistics Toolbox user interface in a web browser.
The user interface can also be accessed by https://statool.shinyapps.io/Chinese. Neither R nor any packages are required in this online version.
However, it is highly recommended that the user interface be launched locally.
}
\examples{
if(interactive()) {
   ChineseUI()
}
}
\author{
Zhicheng Ji <zji4@zji4.edu>
}

