% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EnglishUI.R
\name{EnglishUI}
\alias{EnglishUI}
\title{EnglishUI}
\usage{
EnglishUI()
}
\description{
Launch the Statistics Toolbox (English Version) Graphical User Interface
}
\details{
This function will automatically launch the Statistics Toolbox user interface in a web browser.
The user interface can also be accessed by https://statool.shinyapps.io/English. Neither R nor any packages are required in this online version.
However, it is highly recommended that the user interface be launched locally.
}
\examples{
if(interactive()) {
   EnglishUI()
}
}
\author{
Zhicheng Ji <zji4@zji4.edu>
}

